Change: Move all openGL types to LibRender
