+++ source/OpenBveApi/Routes/StaticBackground.cs
@@ -22,7 +22,7 @@ namespace OpenBveApi.Routes
 		/// <summary> The time at which this background should be displayed (Expressed as the number of seconds since midnight)</summary>
 		public double Time;
 		/// <summary>The OpenGL/OpenTK VAO for the background</summary>
-		public VertexArrayObject VAO;
+		public object VAO;
 
 		/// <summary>Creates a new static background, using the default 0.8s fade-in time</summary>
 		/// <param name="Texture">The texture to apply</param>
@@ -58,122 +58,7 @@ namespace OpenBveApi.Routes
 			this.Mode = Mode;
 			this.Time = Time;
 		}
-
-		public void CreateVAO()
-		{
-			float y0, y1;
-
-			if (KeepAspectRatio)
-			{
-				int tw = Texture.Width;
-				int th = Texture.Height;
-				double hh = System.Math.PI * BackgroundImageDistance * th / (tw * Repetition);
-				y0 = (float)(-0.5 * hh);
-				y1 = (float)(1.5 * hh);
-			}
-			else
-			{
-				y0 = (float)(-0.125 * BackgroundImageDistance);
-				y1 = (float)(0.375 * BackgroundImageDistance);
-			}
-
-			const int n = 32;
-			Vector3f[] bottom = new Vector3f[n];
-			Vector3f[] top = new Vector3f[n];
-			double angleValue = 2.61799387799149 - 3.14159265358979 / n;
-			const double angleIncrement = 6.28318530717958 / n;
-
-			/*
-			 * To ensure that the whole background cylinder is rendered inside the viewing frustum,
-			 * the background is rendered before the scene with z-buffer writes disabled. Then,
-			 * the actual distance from the camera is irrelevant as long as it is inside the frustum.
-			 * */
-			for (int i = 0; i < n; i++)
-			{
-				float x = (float)(BackgroundImageDistance * System.Math.Cos(angleValue));
-				float z = (float)(BackgroundImageDistance * System.Math.Sin(angleValue));
-				bottom[i] = new Vector3f(x, y0, -z);
-				top[i] = new Vector3f(x, y1, -z);
-				angleValue += angleIncrement;
-			}
-
-			float textureStart = 0.5f * (float)Repetition / n;
-			float textureIncrement = -(float)Repetition / n;
-			float textureX = textureStart;
-
-			List<LibRenderVertex> vertexData = new List<LibRenderVertex>();
-			List<int> indexData = new List<int>();
-
-			for (int i = 0; i < n; i++)
-			{
-				int j = (i + 1) % n;
-				int indexOffset = vertexData.Count;
-
-				// side wall
-				vertexData.Add(new LibRenderVertex
-				{
-					Position = top[i],
-					UV = new Vector2f(textureX, 0.005f),
-					Color = Color128.White
-				});
-
-				vertexData.Add(new LibRenderVertex
-				{
-					Position = bottom[i],
-					UV = new Vector2f(textureX, 0.995f),
-					Color = Color128.White
-				});
-
-				vertexData.Add(new LibRenderVertex
-				{
-					Position = bottom[j],
-					UV = new Vector2f(textureX + textureIncrement, 0.995f),
-					Color = Color128.White
-				});
-
-				vertexData.Add(new LibRenderVertex
-				{
-					Position = top[j],
-					UV = new Vector2f(textureX + textureIncrement, 0.005f),
-					Color = Color128.White
-				});
-
-				indexData.AddRange(new[] { 0, 1, 2, 3 }.Select(x => x + indexOffset));
-
-				// top cap
-				vertexData.Add(new LibRenderVertex
-				{
-					Position = new Vector3f(0.0f, top[i].Y, 0.0f),
-					UV = new Vector2f(textureX + 0.5f * textureIncrement, 0.1f),
-					Color = Color128.White
-				});
-
-				indexData.AddRange(new[] { 0, 3, 4 }.Select(x => x + indexOffset));
-
-				// bottom cap
-				vertexData.Add(new LibRenderVertex
-				{
-					Position = new Vector3f(0.0f, bottom[i].Y, 0.0f),
-					UV = new Vector2f(textureX + 0.5f * textureIncrement, 0.9f),
-					Color = Color128.White
-				});
-
-				indexData.AddRange(new[] { 5, 2, 1 }.Select(x => x + indexOffset));
-
-				// finish
-				textureX += textureIncrement;
-			}
-
-			VAO?.UnBind();
-			VAO?.Dispose();
-
-			VAO = new VertexArrayObject();
-			VAO.Bind();
-			VAO.SetVBO(new VertexBufferObject(vertexData.ToArray(), BufferUsageHint.StaticDraw));
-			VAO.SetIBO(new IndexBufferObject(indexData.ToArray(), BufferUsageHint.StaticDraw));
-			VAO.UnBind();
-		}
-
+		
 		public override void UpdateBackground(double SecondsSinceMidnight, double ElapsedTime, bool Target)
 		{
 			if (Target)
